web_tts/requirements.txt
# Web TTS Requirements

# Основные зависимости tts_common
-r ../tts_common/requirements.txt

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Шаблоны
jinja2>=3.1.0

# Для обработки данных из HTML-форм в FastAPI
python-multipart

# Database ORM
sqlalchemy>=2.0.0

# Google Drive API
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Environment variables
python-dotenv>=1.0.0